requirements.txt
pandas
python-bcb
python-dotenv

# google api
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
lxml